requirements.txt
mysql-connector-python==9.0.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7
click==8.1.6
colorama==0.4.6
importlib-metadata==6.6.0
zipp==3.15.0
